Pandora::Gtk classes extracted
