+++ lib/pandora/gtk/panobj_scroll_win.rb
@@ -1,0 +1,10 @@
+module Pandora
+  module Gtk
+
+    # ScrolledWindow for panobjects
+    # RU: ScrolledWindow для объектов Пандоры
+    class PanobjScrollWin < ::Gtk::ScrolledWindow
+    end
+
+  end
+end
